Leitura do excel , passar para uma estrutura de dados

+++ ES1-2019-EIPL-83/src/project/GUI/Method.java
@@ -2,20 +2,20 @@ package project.GUI;
 
 public class Method {
 
-	private int id;
+	private double id;
 	private String inPackage;
 	private String inClass;
 	private String methodName;
-	private int LOC;
-	private int CYCLO;
-	private int ATFD;
+	private double LOC;
+	private double CYCLO;
+	private double ATFD;
 	private float LAA;
 	private boolean is_long_method;
 	private boolean iPlasma;
 	private boolean PMD;
 	private boolean is_feature_envy;
 
-	public Method(int id, String inPackage, String inClass, String methodName, int LOC, int CYCLO, int ATFD, float LAA,
+	public Method(double id, String inPackage, String inClass, String methodName, double LOC, double CYCLO, double ATFD, float LAA,
 			boolean is_long_method, boolean iPlasma, boolean PMD, boolean is_feature_envy) {
 		this.id = id;
 		this.inPackage = inPackage;
